Add GAN-G (not working)
The control parameter is regulated through the suggested error between
current control input and the best candidate calculated using generative
model of pretrained GAN.

The result was not good. It fluctuated so much.
The reason is probably the linear structure of the Q-function,
which continuously produces unstable best candidates

One way to overcome this situation is to let another Q-function
using deep structure to closely model the exact Q-function
of the nonlinear problem.